+++ server/media-service/src/main/java/com/metawebthree/media/domain/model/ArtWorkTagDO.java
@@ -16,7 +16,7 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder
 @TableName("\"Artwork_Tag\"")
 public class ArtWorkTagDO extends BaseDO {
-    @TableId(type=IdType.ASSIGN_ID)
-    Integer id;
+    @TableId(type = IdType.ASSIGN_ID)
+    Long id;
     String tag;
 }
